Extract generator runtime and add coverage
