+++ docs/next-steps.md
@@ -15,6 +15,7 @@
 - Refresh unit/e2e test suites to cover EventBus ordering, CRUD flows, pause/resume backlog, and query-mode lossiness. ✅ Unit coverage now exists for `CDCController`, adapters, metrics widgets, and lane diff schema drift; Playwright exercises the schema walkthrough end-to-end.
 - Update developer docs (onboarding, harness guide) with the new architecture map and event bus workflow. ✅ Added event-flow overview and harness mapping notes.
 - Prepare initial metrics telemetry hooks (in-memory only) and dashboard UI shell. ✅ Dashboard surfaces per-lane metrics with schema walkthrough controls in place.
+- Harden the high-volume generator + replay tooling. ✅ Extracted generator runtime into a tested shared module powering the comparator burst controls and verified workspace replay wiring.
 
 ## Definition of Ready
 - Implementation plan captured at `docs/IMPLEMENTATION_PLAN.md` and linked in team tooling. ✅
